+++ soc/arm/nordic_nrf/common/soc_nrf_common.h
