+++ soc/arm/nordic_nrf/nrf51/soc.h
@@ -11,7 +11,7 @@
 #ifndef _NORDICSEMI_NRF51_SOC_H_
 #define _NORDICSEMI_NRF51_SOC_H_
 
-#include "../soc_nrf_common.h"
+#include <soc_nrf_common.h>
 
 #define NRF51_POWER_RAMON_ADDRESS              0x40000524
 #define NRF51_POWER_RAMONB_ADDRESS             0x40000554
