+++ soc/arm/nordic_nrf/nrf91/soc.h
@@ -11,7 +11,7 @@
 #ifndef _NORDICSEMI_NRF91_SOC_H_
 #define _NORDICSEMI_NRF91_SOC_H_
 
-#include "../soc_nrf_common.h"
+#include <soc_nrf_common.h>
 
 #define FLASH_PAGE_ERASE_MAX_TIME_US	89700UL
 #define FLASH_PAGE_MAX_CNT		256UL
